% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_values.R
\name{get_max_values}
\alias{get_max_values}
\title{Calculate Maximum Spirometry Values by Test}
\usage{
get_max_values(df, grp_vars = NULL, basic_parameters = TRUE)
}
\arguments{
\item{df}{a data frame created by \code{read_spiro_sql()} or \verb{read_spiro_pdf().}}

\item{grp_vars}{bldj}

\item{basic_parameters}{logical; if TRUE, the default, only values for the following parameters are returned:
\itemize{
\item FEV1
\item FVC
\item FEV6
\item FEV1/FVC
\item FEF2575
\item Peak Flow (PEF)
}}
}
\value{
a tibble.
}
\description{
Calculate Maximum Spirometry Values by Test
}
\examples{

\dontrun{
get_max_values(df, test_threshold = NULL, test_col = 'Spiro_Event', grp_vars = NULL,
               basic_parameters = TRUE)
}

}
