% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_summary_values.R
\name{get_summary_values}
\alias{get_summary_values}
\title{Calculate Summary Spirometry Values by Test}
\usage{
get_summary_values(
  df,
  grp_vars = NULL,
  sum_fun = "max",
  basic_parameters = TRUE
)
}
\arguments{
\item{df}{a data frame created by \code{read_spiro_sql()} or \verb{read_spiro_pdf().}}

\item{grp_vars}{character; string of column names to used to calculate maximum values by group (i.e. 'id', 'spiro_test', etc.).
See \link[dplyr]{group_by}.}

\item{sum_fun}{character; summary function to be applied.  Must be one of c("max," "mean," "sd," "median," "min").}

\item{basic_parameters}{logical; if TRUE, the default, only values for the following parameters are returned:
\itemize{
\item FEV1
\item FVC
\item FEV6
\item FEV1/FVC
\item FEF2575
\item Peak Flow (PEF)
}}
}
\description{
\code{get_summary_values} calculates summary statistics for each parameter for all trials within a test.
The intent of this function is to find the maximum value of each parameter as is done according to American
Thoracic society criteria, but other summary statistics can be generated as well (mean, median, sd).
Users can also specify grouping variables, and table operations will be performed 'by group.'
}
\examples{

\dontrun{
get_max_values(df, test_threshold = NULL, grp_vars = c('ID', 'spiro_test'), sum_fun = 'max',
               basic_parameters = TRUE)
}

}
