% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spiro_sql.R
\name{read_spiro_sql}
\alias{read_spiro_sql}
\title{Read Spirobank SQLite database}
\usage{
read_spiro_sql(
  sqldb = NULL,
  demo = FALSE,
  inspiratory = FALSE,
  tzone = "America/New_York",
  participant_id = NULL,
  sample_col = NULL,
  sample_id = NULL
)
}
\arguments{
\item{sqldb}{character; path of Spirobank SQLite database}

\item{demo}{logical; include patient demographics and patient anthropometrics?
Default = FALSE.}

\item{inspiratory}{logical, include inspiratory maneauvers? Default = FALSE.}

\item{tzone}{character, time zone of location where measurments were recorded.
Default = 'America/New_York'.}

\item{participant_id}{user defined string to denote a patient identifier.
If specified, a new column is created ('ID') listing the string. Default is NULL.}

\item{sample_col;}{character; user defined character string specifying the name of the
column to denote a sample ID. This column is useful to denote repeated periods of
data collection by individual patents. Default is NULL.}

\item{sample_id;}{user defined string to denote sample ID. If assigned, a
value must also be supplied to \code{sample_col}. Default is NULL}
}
\value{
a tibble with spirometry values.
}
\description{
Read Spirobank SQLite database
}
\examples{
\dontrun{
read_spiro_sql(sqldb = NULL, demo = FALSE, inspiratory = FALSE, tzone = 'America/New_York',
               participant_id = NULL, sample_col = NULL, sample_id = NULL )
}

}
