% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_max_values.R
\name{get_max_values}
\alias{get_max_values}
\title{Calculate Maximum Spirometry Values by Test}
\usage{
get_max_values(
  df,
  test_threshold = NULL,
  test_col = "Spiro_Event",
  sample_col = NULL,
  participant_id = NULL,
  basic_parameters = TRUE
)
}
\arguments{
\item{df}{a data frame created by \code{read_spiro_sql()} or \verb{read_spiro_pdf().}}

\item{test_threshold}{numeric; threshold value (seconds) used partition data into discrete testing sessions.
If the time difference between the last trial and and a new trial exceeds the threshold value, the current
trial is designated as belonging to a new test. For example, the threshold value is set at 600 seconds (10 minutes)
and a patient performs 6 trials  The first 3 trials are done consecutively within 10 minutes.  The fourth trial occurs
about one hour after the third.  The fifth and sixth trial are completed within 10 minutes of the fourth trial.
These six trials represent two separate testing sessions according the user defined threshold.

If NULL, the default, the Spirobank defined session variable (\code{spiro_session}) will be used.}

\item{test_col}{character; name of column to denote testing sessions if \code{test_threshold} is specified,
otherwise it will be ignored.  Default = 'Spiro_Event.'}

\item{sample_col}{character; name of column to denote patient sample number. Useful if a patient has
more than one period of data collection recorded in the input data frame.}

\item{participant_id}{string; overwrites value of \code{ID} in the input data frame.}

\item{basic_parameters}{logical; if TRUE, the default, only values for the following parameters are returned:
\itemize{
\item FEV1
\item FVC
\item FEV6
\item FEV1/FVC
\item FEF2575
\item Peak Flow (PEF)
}}
}
\value{
a tibble.
}
\description{
Calculate Maximum Spirometry Values by Test
}
\examples{

\dontrun{
get_max_values(df, test_threshold = NULL, test_col = 'Spiro_Event', sample_col = NULL, participant_id = NULL,
               basic_parameters = TRUE)
}

}
