% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spiro_sql.R
\name{read_spiro_sql}
\alias{read_spiro_sql}
\title{Read Spirobank SQLite database}
\usage{
read_spiro_sql(
  sqldb = NULL,
  demo = FALSE,
  inspiratory = FALSE,
  tzone = "America/New_York"
)
}
\arguments{
\item{sqldb}{character; path of Spirobank SQLite database}

\item{demo}{logical; include patient demographics and patient anthropometrics?
Default = FALSE.}

\item{inspiratory}{logical, include inspiratory maneauvers? Default = FALSE.}

\item{tzone}{character, time zone of location where measurments were recorded.
Default = 'America/New_York'.}
}
\value{
a tibble with spirometry values.
}
\description{
Read Spirobank SQLite database
}
\examples{
\dontrun {
read_spiro_sql <- function(sqldb = NULL, demo = FALSE, inspiratory = FALSE, tzone = 'America/New_York')
}

}
